Refactored the integration test code to make it easier to fetch emails sent to the student.
