+++ grader/src/it/java/edu/pdx/cs410J/grader/EmailSenderIntegrationTestCase.java
@@ -1,20 +1,17 @@
 package edu.pdx.cs410J.grader;
 
 import com.icegreen.greenmail.imap.AuthorizationException;
-import com.icegreen.greenmail.imap.ImapHostManager;
 import com.icegreen.greenmail.store.FolderException;
-import com.icegreen.greenmail.store.FolderListener;
-import com.icegreen.greenmail.store.MailFolder;
 import com.icegreen.greenmail.user.GreenMailUser;
 import com.icegreen.greenmail.util.GreenMail;
 import com.icegreen.greenmail.util.ServerSetup;
 import org.junit.After;
 import org.junit.Before;
 
-import javax.mail.Flags;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
+import java.util.List;
 
 public abstract class EmailSenderIntegrationTestCase {
   protected final String emailServerHost = "127.0.0.1";
@@ -24,7 +21,6 @@ public abstract class EmailSenderIntegrationTestCase {
   protected final int imapsPort = 9933;
   protected final File tempDirectory = new File(System.getProperty("java.io.tmpdir"));
   protected GreenMail emailServer;
-  private String graderEmail = EmailSender.TA_EMAIL.getAddress();
 
   @Before
   public void startEmailServer() throws FolderException, AuthorizationException {
@@ -32,44 +28,18 @@ public abstract class EmailSenderIntegrationTestCase {
     ServerSetup imaps = new ServerSetup(imapsPort, emailServerHost, ServerSetup.PROTOCOL_IMAPS);
     emailServer = new GreenMail(new ServerSetup[]{ smtp, imaps });
 
-    GreenMailUser graderUser = emailServer.setUser(graderEmail, imapUserName, imapPassword);
-
-    moveEmailsFromInboxToProjectSubmissions(graderUser);
+    List<String> addresses = getEmailAddressesForSmtpServer();
+    for (String address : addresses) {
+      GreenMailUser user = emailServer.setUser(address, imapUserName, imapPassword);
+      initializeSmtpUser(user);
+    }
 
     emailServer.start();
   }
 
-  protected void moveEmailsFromInboxToProjectSubmissions(GreenMailUser user) throws AuthorizationException, FolderException {
-    ImapHostManager manager = emailServer.getManagers().getImapHostManager();
-    MailFolder submissions = manager.createMailbox(user, getEmailFolderName());
-    MailFolder inbox = manager.getInbox(user);
-    inbox.addListener(new FolderListener() {
-      @Override
-      public void expunged(int msn) {
+  protected abstract void initializeSmtpUser(GreenMailUser user) throws AuthorizationException, FolderException;
 
-      }
-
-      @Override
-      public void added(int msn) {
-        try {
-          inbox.copyMessage(msn, submissions);
-
-        } catch (FolderException ex) {
-          throw new IllegalStateException("Can't copy message to submissions folder", ex);
-        }
-      }
-
-      @Override
-      public void flagsUpdated(int msn, Flags flags, Long uid) {
-
-      }
-
-      @Override
-      public void mailboxDeleted() {
-
-      }
-    });
-  }
+  protected abstract List<String> getEmailAddressesForSmtpServer();
 
   @After
   public void stopEmailServer() {
@@ -93,5 +63,4 @@ public abstract class EmailSenderIntegrationTestCase {
     return dir;
   }
 
-  protected abstract String getEmailFolderName();
 }
